a lot of fixes big split changes

+++ Projects/Libft/ft_strncmp.c
@@ -6,15 +6,15 @@
 /*   By: hariskon <hariskon@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/23 17:15:11 by hariskon          #+#    #+#             */
-/*   Updated: 2025/05/05 15:19:18 by hariskon         ###   ########.fr       */
+/*   Updated: 2025/05/09 14:38:53 by hariskon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strncmp(const char *s1, const char *s2, int size)
+int	ft_strncmp(const char *s1, const char *s2, size_t size)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while ((s1[i] || s2[i]) && i < size)
